Download environment from dachuer/vcpkg

+++ tools/macos_buildenv.sh
@@ -40,21 +40,27 @@ case "$1" in
         if [ ! -d "${BUILDENV_PATH}" ]; then
             if [ "$1" != "--profile" ]; then
                 echo "Build environment $BUILDENV_NAME not found in mixxx repository, downloading it..."
-                curl "https://downloads.mixxx.org/dependencies/2.4/macOS/${BUILDENV_NAME}.zip" -o "${BUILDENV_PATH}.zip"
-                OBSERVED_SHA256=$(shasum -a 256 "${BUILDENV_PATH}.zip"|cut -f 1 -d' ')
-                if [[ "$OBSERVED_SHA256" == "$BUILDENV_SHA256" ]]; then
-                    echo "Download matched expected SHA256 sum $BUILDENV_SHA256"
-                else
-                    echo "ERROR: Download did not match expected SHA256 checksum!"
+                if curl -o "${BUILDENV_PATH}_.zip" -L -H "authorization: token $2" -H "Accept: application/vnd.github.v3+json" https://api.github.com/repos/daschuer/vcpkg/actions/artifacts/229834703/zip; then
+                    #OBSERVED_SHA256=$(shasum -a 256 "${BUILDENV_PATH}.zip"|cut -f 1 -d' ')
+                    #if [[ "$OBSERVED_SHA256" == "$BUILDENV_SHA256" ]]; then
+                    #    echo "Download matched expected SHA256 sum $BUILDENV_SHA256"
+                    #else
+                    #    echo "ERROR: Download did not match expected SHA256 checksum!"
                     echo "Expected $BUILDENV_SHA256"
-                    echo "Got $OBSERVED_SHA256"
-                    exit 1
+                    #    echo "Got $OBSERVED_SHA256"
+                    #    exit 1
+                    #fi
+                    echo "Extracting ${BUILDENV_NAME}_.zip..."
+                    unzip "${BUILDENV_PATH}_.zip" -d "${BUILDENV_BASEPATH}" && \
+                    echo "Successfully extracted ${BUILDENV_NAME}_.zip" && \
+                    rm "${BUILDENV_PATH}_.zip"
+                    echo "Extracting ${BUILDENV_NAME}.zip..."
+                    unzip "${BUILDENV_PATH}.zip" -d "${BUILDENV_BASEPATH}" && \
+                    echo "Successfully extracted ${BUILDENV_NAME}.zip" && \
+                    rm "${BUILDENV_PATH}.zip"
+                else
+                   echo "Downloading build environment failed"
                 fi
-                echo ""
-                echo "Extracting ${BUILDENV_NAME}.zip..."
-                unzip "${BUILDENV_PATH}.zip" -d "${BUILDENV_BASEPATH}" && \
-                echo "Successfully extracted ${BUILDENV_NAME}.zip" && \
-                rm "${BUILDENV_PATH}.zip"
             else
                 echo "Build environment $BUILDENV_NAME not found in mixxx repository, run the command below to download it."
                 echo "source ${THIS_SCRIPT_NAME} setup"
